- fix bug when updating parameters with index > 255 (sliders on UI not updating)

+++ hi_core/hi_core/MacroControlledComponents.h
@@ -367,7 +367,7 @@ public:
 	void setModulationData(ModulationPopupData::Ptr modData);
 
 
-	void onAttributeChange(dispatch::library::Processor* p, uint8 index);
+	void onAttributeChange(dispatch::library::Processor* p, uint16 index);
 
 	juce::AudioProcessorParameter* getConnectedPluginParameter() const;
